docker-compose: mount the database folder instead of the file
When mounting the database.sqlite file while it does not exist, docker
creates instead a new folder called 'database.sqlite'.

To fix it, mount the database folder instead, so that the application
can generate the 'database.sqlite' file in the '/app/database' folder.

Note that the '.env' file has a similar problem, which cannot be fixed
in the same way (because we would have to mount the entire top-level
folder as a volume).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>