README.md: document populating .env before using docker-compose
Currently if docker-compose is run before a .env file is created, a
folder will be created instead which leads to confusion about what
should be done.

Since we don't have a fix for it yet, just document in the README file
that the .env file should be created first.

While at it, also mention the APP_KEY which is also mandatory.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>